te-3767: creating separate rules for checking structure of twig, scss and typescript files
